# docs/swagger.yaml
definitions:
  handler.CreateCompanyRequest:
    properties:
      address:
        type: string
      addressCity:
        type: string
      addressComplement:
        type: string
      addressNumber:
        type: string
      addressState:
        type: string
      addressZipCode:
        type: string
      cnpj:
        type: string
      name:
        type: string
    type: object
  handler.CreateCompanyResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.CompanyResponse'
      message:
        type: string
    type: object
  handler.DeleteCompanyResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.CompanyResponse'
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  handler.ListCompaniesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.CompanyResponse'
        type: array
      message:
        type: string
    type: object
  handler.ShowCompanyResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.CompanyResponse'
      message:
        type: string
    type: object
  handler.UpdateCompanyRequest:
    properties:
      address:
        type: string
      addressCity:
        type: string
      addressComplement:
        type: string
      addressNumber:
        type: string
      addressState:
        type: string
      addressZipCode:
        type: string
      cnpj:
        type: string
      name:
        type: string
    type: object
  handler.UpdateCompanyResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.CompanyResponse'
      message:
        type: string
    type: object
  schemas.CompanyResponse:
    properties:
      address:
        type: string
      addressCity:
        type: string
      addressComplement:
        type: string
      addressNumber:
        type: string
      addressState:
        type: string
      addressZipCode:
        type: string
      cnpj:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/schemas.UserResponse'
        type: array
    type: object
  schemas.UserResponse:
    properties:
      address:
        type: string
      companyId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /companies:
    get:
      consumes:
      - application/json
      description: List all Companies
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListCompaniesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Companies
      tags:
      - Company
  /company:
    delete:
      consumes:
      - application/json
      description: Delete a company
      parameters:
      - description: Company identification
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteCompanyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Company
      tags:
      - Company
    get:
      consumes:
      - application/json
      description: Show a company
      parameters:
      - description: Company identification
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShowCompanyResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Show Company
      tags:
      - Company
    post:
      consumes:
      - application/json
      description: Create a new company
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCompanyRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateCompanyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Company
      tags:
      - Company
    put:
      consumes:
      - application/json
      description: Update a new company
      parameters:
      - description: Company identification
        in: query
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCompanyRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.UpdateCompanyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Company
      tags:
      - Company
swagger: "2.0"